+++ NEWS.md
@@ -6,6 +6,7 @@
 * some reusable regular expressions
 * some new functions to work with multiple seqtab objects (`seqtab_stats_bind` and `seqtab_stats_plot`)
 * better feature naming options for `ft` objects
+* new `obs_callout()` function
 
 # jakR2 0.2